move to stream pkf wip
