+++ lib/cchain/provider/provider.go
@@ -72,7 +72,6 @@ type Provider struct {
 	backoffFunc func(context.Context) func()
 	chainNamer  func(xchain.ChainVersion) string
 	network     netconf.ID
-	cache       *attestationCache
 }
 
 // NewProviderForT creates a new provider for testing.
@@ -85,7 +84,6 @@ func NewProviderForT(_ *testing.T, fetch fetchFunc, latest latestFunc, window wi
 		window:      window,
 		backoffFunc: backoffFunc,
 		chainNamer:  func(xchain.ChainVersion) string { return "" },
-		cache:       newAttestationCache(),
 	}
 }
 
@@ -182,23 +180,8 @@ func (p Provider) stream(
 	ctx := log.WithCtx(in, "src_chain", srcChain, "worker", workerName)
 
 	deps := stream.Deps[xchain.Attestation]{
-		FetchBatch: func(ctx context.Context, _ uint64, offset uint64) ([]xchain.Attestation, error) {
-			// If cached attestations found, return them.
-			cachedAtts := p.cache.get(chainVer, offset)
-			if len(cachedAtts) > 0 {
-				return cachedAtts, nil
-			}
-
-			// If the height is not cached, fetch the attestations.
-			atts, err := p.fetch(ctx, chainVer, offset)
-			if err != nil {
-				return nil, err
-			}
-
-			// Instruct the cache to be updated.
-			p.cache.update(chainVer, atts)
-
-			return atts, nil
+		FetchBatch: func(ctx context.Context, chainVer xchain.ChainVersion, offset uint64) ([]xchain.Attestation, error) {
+			return p.fetch(ctx, chainVer, offset)
 		},
 		Backoff:       p.backoffFunc,
 		ElemLabel:     "attestation",
@@ -241,7 +224,7 @@ func (p Provider) stream(
 	}
 
 	cb := (stream.Callback[xchain.Attestation])(callback)
-	err := stream.Stream(ctx, deps, chainVer.ID, attestOffset, cb)
+	err := stream.Stream(ctx, deps, chainVer, attestOffset, cb)
 	if err != nil {
 		return errors.Wrap(err, "stream attestations", "worker", workerName, "chain", srcChain)
 	}
