+++ lib/stream/cache.go
@@ -1,4 +1,4 @@
-package provider
+package stream
 
 import (
 	"math"
